move capture_output util from parallel tests to utils.io
